+++ slices/sorted_index_test.go
@@ -81,6 +81,18 @@ func TestSortedIndex(t *testing.T) {
 			value: 3,
 			want:  3,
 		},
+		{
+			name:  "more more sequence",
+			arr:   []int{0, 1, 2, 3, 3, 3, 3, 3, 3, 4, 5, 6, 7, 8, 9},
+			value: 3,
+			want:  3,
+		},
+		{
+			name:  "does not exist",
+			arr:   []int{0, 1, 2, 3, 4, 6, 7, 8, 9},
+			value: 10,
+			want:  -1,
+		},
 	}
 
 	for _, subject := range tests {